fix(clang-tidy): address all 17 bugprone/performance findings
Five categories from the first clang-tidy run, all real:

- bugprone-narrowing-conversions (int64_t → double in throughput math):
  oms_demo / risk_demo / logger_demo / strategy_demo each had
  'total_ns / 1e6' implicitly narrowing. Wrap in static_cast<double>().
- bugprone-narrowing-conversions (uint32 → int32 on ITCH shares field):
  test_all.cpp line 968 — explicit static_cast<int32_t>().
- bugprone-implicit-widening-of-multiplication-result:
  strategy_demo:203 'i * 1000' into int64_t param   → cast i first.
  test_all.cpp:601 'PRODS * PER / CONS * 2' for size_t reserve → cast
  PRODS to std::size_t first so the whole expression widens.
- bugprone-not-null-terminated-result on ITCH/OUCH wire-format symbol
  fields (fixed-width, intentionally not null-terminated):
  add NOLINTNEXTLINE with a comment so future readers know it's
  deliberate, not an oversight.
- performance-inefficient-vector-operation: four emplace_back loops
  (test_mpsc, test_mpmc producers + consumers, test_sequencer
  consumers) — call reserve(N) before the loop.

Flip clang-tidy back to blocking — no remaining warnings to triage.

+++ logger/logger_demo.cpp
@@ -192,10 +192,10 @@ void benchmark(int num_events) {
     std::sort(latencies.begin(), latencies.end());
     int n = static_cast<int>(latencies.size());
     double avg = static_cast<double>(total_ns) / n;
-    double throughput = n / (total_ns / 1e9) / 1e6;
+    double throughput = n / (static_cast<double>(total_ns) / 1e9) / 1e6;
 
     printf("--- Results ---\n");
-    printf("  Total time:  %.2f ms\n", total_ns / 1e6);
+    printf("  Total time:  %.2f ms\n", static_cast<double>(total_ns) / 1e6);
     printf("  Avg:         %.0f ns/event\n", avg);
     printf("  p50:         %ld ns\n", latencies[n / 2]);
     printf("  p90:         %ld ns\n", latencies[(int)(n * 0.90)]);
